some debugging - not working yet

+++ artist/physics_objects/heliostats/concentrator/facets/point_cloud_facets.py
@@ -53,8 +53,10 @@ class PointCloudFacetModule(AFacetModule):
             )
         else:
             self.surface_points = torch.tensor(
-                config_file["heliostats"]["general_surface_points"][()]
+                config_file["heliostats"]["general_surface_points"][()],
+                dtype=torch.float,
             )
             self.surface_normals = torch.tensor(
-                config_file["heliostats"]["general_surface_normals"][()]
+                config_file["heliostats"]["general_surface_normals"][()],
+                dtype=torch.float,
             )
